✨ refactor: Update imports for EditorGroupTabLabel
Refactor the import statements across multiple files to 
use the correct package for `EditorGroupTabLabel`. This 
change improves code organization and clarity by ensuring 
that the label class is consistently from its 
design package.
