+++ src/main/java/krasa/editorGroups/tabs2/impl/KrActionPanel.java
@@ -13,6 +13,7 @@ import com.intellij.ui.components.panels.NonOpaquePanel;
 import com.intellij.util.ui.UIUtil;
 import krasa.editorGroups.tabs2.EditorGroupsPanelTabs;
 import krasa.editorGroups.tabs2.label.EditorGroupTabInfo;
+import krasa.editorGroups.tabs2.label.EditorGroupTabLabel;
 
 import javax.swing.*;
 import java.awt.*;
